test-related Gulp tasks refactored to a separate file
