+++ package.json
@@ -21,7 +21,8 @@
     "package": "gulp package",
     "test": "gulp test",
     "unit-test": "gulp unit-test",
-    "browser-test": "gulp browser-test"
+    "browser-test": "gulp browser-test",
+    "ui-test": "gulp ui-test"
   },
   "dependencies": {
     "fs-extra": "0.26.7",
